# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gateway microservice of youtube-clone app.",
        "title": "rest api of gateway service",
        "contact": {
            "name": "rzaf",
            "url": "https://www.github.com/rzaf/youtube-clone"
        }
    },
    "basePath": "/api",
    "paths": {
...
        "/medias": {
            "get": {
                "description": "get medias",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "get medias",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "items perpage",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "media creator",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "photo",
                            "video",
                            "music",
                            "any"
                        ],
                        "type": "string",
                        "default": "video",
                        "description": "media type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "oldest",
                            "most-viewed",
                            "least-viewed"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "sort type",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create media",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "create media",
                "parameters": [
                    {
                        "enum": [
                            "photo",
                            "video",
                            "music"
                        ],
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/medias/search/{term}": {
            "get": {
                "description": "search medias",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "search medias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search term",
                        "name": "term",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "items perpage",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "media creator",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "photo",
                            "video",
                            "music",
                            "any"
                        ],
                        "type": "string",
                        "default": "video",
                        "description": "media type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "oldest",
                            "most-viewed",
                            "least-viewed"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "sort type",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/medias/{url}/tag/{name}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add tag to media",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "add tag to media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "remove tag from media",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "remove tag from media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/playlists/{url}/medias": {
            "get": {
                "description": "get medias of playlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medias"
                ],
                "summary": "get medias of playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "items perpage",
                        "name": "perpage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "request failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "api key of user",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}